Move AppBootstrapper registration to separate scene BootstrapInstaller
While AppBootstrapper in ProjectInstaller it will run game fsm & unload InitTestScene, so play mode tests won't run
