+++ src/DungeonBrothers/Assets/CodeBase/Infrastructure/Installers/BootstrapInstaller.cs
@@ -1,0 +1,12 @@
+﻿using CodeBase.Infrastructure.StateMachines.App;
+using VContainer;
+using VContainer.Unity;
+
+namespace CodeBase.Infrastructure.Installers
+{
+    public class BootstrapInstaller : LifetimeScope
+    {
+        protected override void Configure(IContainerBuilder builder) => 
+            builder.Register<AppBootstrapper>(Lifetime.Singleton).AsImplementedInterfaces();
+    }
+}
