+++ src/DungeonBrothers/Assets/CodeBase/Infrastructure/Installers/BootstrapInstaller.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 4b9477e038204ec8b89eaa4d557b5f3c
+timeCreated: 1696185442
