+++ src/DungeonBrothers/Assets/CodeBase/Infrastructure/Installers/ProjectInstaller.cs
@@ -6,7 +6,6 @@ using CodeBase.Infrastructure.Services.Providers.CameraProvider;
 using CodeBase.Infrastructure.Services.Providers.CharactersProvider;
 using CodeBase.Infrastructure.Services.SceneLoader;
 using CodeBase.Infrastructure.Services.StaticDataProvider;
-using CodeBase.Infrastructure.StateMachines.App;
 using CodeBase.Infrastructure.StateMachines.App.FSM;
 using CodeBase.Infrastructure.StateMachines.App.States;
 using CodeBase.Infrastructure.StateMachines.Gameplay.States;
@@ -30,7 +29,6 @@ namespace CodeBase.Infrastructure.Installers
 
         private void RegisterStateMachine(IContainerBuilder builder)
         {
-            builder.Register<AppBootstrapper>(Lifetime.Singleton).AsImplementedInterfaces();
             builder.Register<IAppStateMachine, AppStateMachine>(Lifetime.Singleton);
             
             builder.Register<InitializationState>(Lifetime.Singleton);
