+++ src/DungeonBrothers/Assets/CodeBase/Tests/PlayMode/ExamplePlayModeTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6ac401540bd4f874fadfbeacee3fdb07
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
