+++ src/DungeonBrothers/Assets/CodeBase/Tests/PlayMode/PlayModeTests.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5f9e8d6c773e5bf46a9026d83e2608e5
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
